test(sagas): share the blob saga concurrency tests with other conflict-detecting stores
The five ETag tests in BlobSagaStorageTests describe the ConcurrencyStamp contract rather than Blob Storage, so any store that detects conflicts should run them. They move into a ConcurrentSagaStoreTests fixture beside the base contract; Blob inherits it now and the Redis store follows once it populates the stamp. Only the stale-stamp literal changed, from "etag" to "stale".
